requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0

# Segment Anything Model
segment-anything>=1.0
git+https://github.com/facebookresearch/segment-anything.git

# Medical Image Processing
nibabel>=5.1.0
SimpleITK>=2.3.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
scipy>=1.11.0
albumentations>=1.3.0

# Evaluation Metrics
scikit-learn>=1.3.0
medpy>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Streamlit & Web
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
streamlit-option-menu>=0.3.6

# LLM & RAG Integration
openai>=1.3.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.16.0

# Explainability
grad-cam>=1.4.0
captum>=0.6.0

# Additional
einops>=0.7.0
timm>=0.9.0
